recommand&report

+++ api.js
@@ -456,43 +456,43 @@ app.put('/api/posting/:postingid', authenticateToken, upload.array('photo', 5), 
     const userid = req.user;
     const uploaddate = new Date();
 
-    const updatePostingSql = 'UPDATE posting SET disclosure=?, content=?, roadname=? WHERE postingid=?;';
-    const selectPictureSql = 'SELECT * FROM picture WHERE postingid=?;';
-    const deletePictureSql = 'DELETE FROM picture WHERE postingid=?;';
-    const insertPictureSql = 'INSERT INTO picture (userid, pictureid, name, date, extension, postingid) VALUES (?, ?, ?, ?, ?, ?)';
-
-    db.get().getConnection((err, connection) => { // 커넥션 가져오기
+    // 게시물의 현재 disclosure 상태를 확인합니다.
+    const checkDisclosureSql = 'SELECT disclosure FROM posting WHERE postingid = ?;';
+    db.get().query(checkDisclosureSql, postingid, (err, result) => {
         if (err) {
             console.error(err);
-            return res.status(500).json({ message: '내부 서버 오류' });
+            const error = { "errorCode": "U022", "message": "게시물의 disclosure 상태를 확인하는 동안 오류가 발생했습니다." };
+            return res.status(500).json(error);
         }
 
-        connection.beginTransaction((err) => { // 트랜잭션 시작
+        const currentDisclosure = result[0].disclosure;
+
+        if (currentDisclosure == "비공개(신고당한 게시물)") {
+            const error = { "errorCode": "U023", "message": "비공개(신고당한 게시물) 상태인 게시물은 수정할 수 없습니다." };
+            return res.status(400).json(error);
+        }
+
+        const updatePostingSql = 'UPDATE posting SET disclosure=?, content=?, roadname=? WHERE postingid=?;';
+        const selectPictureSql = 'SELECT * FROM picture WHERE postingid=?;';
+        const deletePictureSql = 'DELETE FROM picture WHERE postingid=?;';
+        const insertPictureSql = 'INSERT INTO picture (userid, pictureid, name, date, extension, postingid) VALUES (?, ?, ?, ?, ?, ?)';
+
+        db.get().getConnection((err, connection) => { // 커넥션 가져오기
             if (err) {
                 console.error(err);
-                connection.release(); // 커넥션 반환
                 return res.status(500).json({ message: '내부 서버 오류' });
             }
 
-            // 게시물 조회
-            const selectPostingSql = "SELECT * FROM posting WHERE postingid = ?";
-            connection.query(selectPostingSql, [postingid, userid], (err, result) => {
+            connection.beginTransaction((err) => { // 트랜잭션 시작
                 if (err) {
                     console.error(err);
-                    connection.rollback(() => {
-                        connection.release(); // 커넥션 반환
-                        return res.status(500).json({ message: '내부 서버 오류' });
-                    });
-                }
-
-                if (result.length === 0) { // 글이 존재하지 않는 경우
-                    const error = { "errorCode": "U010", "message": "핀을 찾을 수 없습니다." };
                     connection.release(); // 커넥션 반환
-                    return res.status(404).json(error);
+                    return res.status(500).json({ message: '내부 서버 오류' });
                 }
 
-                // 게시물 수정
-                connection.query(updatePostingSql, [disclosure, content, roadname, postingid], (err, result) => {
+                // 게시물 조회
+                const selectPostingSql = "SELECT * FROM posting WHERE postingid = ?";
+                connection.query(selectPostingSql, [postingid, userid], (err, result) => {
                     if (err) {
                         console.error(err);
                         connection.rollback(() => {
@@ -500,24 +500,15 @@ app.put('/api/posting/:postingid', authenticateToken, upload.array('photo', 5), 
                             return res.status(500).json({ message: '내부 서버 오류' });
                         });
                     }
-                    
 
-                    //서버에서 기존 사진 삭제
-                    connection.query(selectPictureSql, postingid, (err, result) => {
-                        if(result.length > 0){
-                            result.forEach(function(picture){
-                                file = './pictures/' + picture.pictureid;
-                                fs.unlink(file, function(err){
-                                    if(err) {
-                                        console.log(err);
-                                    }
-                                });
-                            });
-                        }
-                    });
+                    if (result.length === 0) { // 글이 존재하지 않는 경우
+                        const error = { "errorCode": "U010", "message": "핀을 찾을 수 없습니다." };
+                        connection.release(); // 커넥션 반환
+                        return res.status(404).json(error);
+                    }
 
-                    // 데이터베이스에서 기존 사진 삭제
-                    connection.query(deletePictureSql, [postingid], (err, result) => {
+                    // 게시물 수정
+                    connection.query(updatePostingSql, [disclosure, content, roadname, postingid], (err, result) => {
                         if (err) {
                             console.error(err);
                             connection.rollback(() => {
@@ -525,51 +516,79 @@ app.put('/api/posting/:postingid', authenticateToken, upload.array('photo', 5), 
                                 return res.status(500).json({ message: '내부 서버 오류' });
                             });
                         }
+                    
 
-                        // 새로운 사진 추가
-                        const picturePromises = files.map(file => {
-                            const originalname = file.originalname;
-                            const pictureid = file.filename;
-                            const extension = file.mimetype;
-                            return new Promise((resolve, reject) => {
-                                connection.query(insertPictureSql, [userid, pictureid, originalname, uploaddate, extension, postingid], (err, result) => {
-                                    if (err) {
-                                        reject(err);
-                                    } else {
-                                        resolve();
-                                    }
+                        //서버에서 기존 사진 삭제
+                        connection.query(selectPictureSql, postingid, (err, result) => {
+                            if(result.length > 0){
+                                result.forEach(function(picture){
+                                    file = './pictures/' + picture.pictureid;
+                                    fs.unlink(file, function(err){
+                                        if(err) {
+                                            console.log(err);
+                                        }
+                                    });
                                 });
-                            });
+                            }
                         });
 
-                        Promise.all(picturePromises)
-                            .then(() => {
-                                connection.commit((err) => { // 트랜잭션 커밋
-                                    if (err) {
-                                        console.error(err);
-                                        connection.rollback(() => {
-                                            connection.release(); // 커넥션 반환
-                                            return res.status(500).json({ message: '내부 서버 오류' });
-                                        });
-                                    } else {
-                                        connection.release(); // 커넥션 반환
-                                        return res.status(200).json({ message: '게시물이 수정되었습니다.' });
-                                    }
-                                });
-                            })
-                            .catch(err => {
+                        // 데이터베이스에서 기존 사진 삭제
+                        connection.query(deletePictureSql, [postingid], (err, result) => {
+                            if (err) {
                                 console.error(err);
                                 connection.rollback(() => {
                                     connection.release(); // 커넥션 반환
                                     return res.status(500).json({ message: '내부 서버 오류' });
                                 });
+                            }
+
+                            // 새로운 사진 추가
+                            const picturePromises = files.map(file => {
+                                const originalname = file.originalname;
+                                const pictureid = file.filename;
+                                const extension = file.mimetype;
+                                return new Promise((resolve, reject) => {
+                                    connection.query(insertPictureSql, [userid, pictureid, originalname, uploaddate, extension, postingid], (err, result) => {
+                                        if (err) {
+                                            reject(err);
+                                        } else {
+                                            resolve();
+                                        }
+                                    });
+                                });
                             });
+
+                            Promise.all(picturePromises)
+                                .then(() => {
+                                    connection.commit((err) => { // 트랜잭션 커밋
+                                        if (err) {
+                                            console.error(err);
+                                            connection.rollback(() => {
+                                                connection.release(); // 커넥션 반환
+                                                return res.status(500).json({ message: '내부 서버 오류' });
+                                            });
+                                        } else {
+                                            connection.release(); // 커넥션 반환
+                                            return res.status(200).json({ message: '게시물이 수정되었습니다.' });
+                                        }
+                                    });
+                                })
+                                .catch(err => {
+                                    console.error(err);
+                                    connection.rollback(() => {
+                                        connection.release(); // 커넥션 반환
+                                        return res.status(500).json({ message: '내부 서버 오류' });
+                                    });
+                                });
+                        });
                     });
                 });
             });
         });
     });
 });
+
+
   
 // 글 삭제
 app.delete('/api/posting/:postingid', authenticateToken, (req, res) => {
@@ -662,6 +681,136 @@ app.delete('/api/posting/:postingid', authenticateToken, (req, res) => {
                 });
             });
         });
+    });
+});
+
+// 클라이언트가 '추천' 버튼을 눌렀을 때 실행되는 핸들러
+app.post('/api/recommend', authenticateToken, (req, res) => {
+    const { postingid } = req.body;
+    const userid = req.user;
+
+    if (cancel) {
+        // 추천 취소를 요청한 경우
+        const deleteSql = 'DELETE FROM recommand WHERE userid = ? AND postingid = ?;';
+        db.get().query(deleteSql, [userid, postingid], (err) => {
+            if (err) {
+                console.error(err);
+                const error = { "errorCode": "U016", "message": "추천을 취소하는 동안 오류가 발생했습니다." };
+                return res.status(500).json(error);
+            }
+
+            // 현재 게시글의 추천 수를 가져옵니다.
+            const countSql = 'SELECT COUNT(*) AS count FROM recommand WHERE postingid = ?;';
+            db.get().query(countSql, postingid, (err, result) => {
+                if (err) {
+                    console.error(err);
+                    const error = { "errorCode": "U015", "message": "추천 수를 가져오는 동안 오류가 발생했습니다." };
+                    return res.status(500).json(error);
+                }
+
+                const count = result[0].count;
+                res.json({ "count": count });
+            });
+        });
+    } else {
+        // 추천 추가를 요청한 경우
+        const checkSql = 'SELECT * FROM recommand WHERE userid = ? AND postingid = ? LIMIT 1;';
+        db.get().query(checkSql, [userid, postingid], (err, rows) => {
+            if (err) {
+                console.error(err);
+                const error = { "errorCode": "U012", "message": "추천을 확인하는 동안 오류가 발생했습니다." };
+                return res.status(500).json(error);
+            }
+
+            if (rows.length > 0) {
+                // 이미 추천이 존재하는 경우
+                const error = { "errorCode": "U013", "message": "이미 추천한 게시글입니다." };
+                return res.status(400).json(error);
+            } else {
+                // 추천을 추가합니다.
+                const insertSql = 'INSERT INTO recommand (userid, postingid) VALUES (?, ?);';
+                db.get().query(insertSql, [userid, postingid], (err) => {
+                    if (err) {
+                        console.error(err);
+                        const error = { "errorCode": "U014", "message": "추천을 등록하는 동안 오류가 발생했습니다." };
+                        return res.status(500).json(error);
+                    }
+                    // 현재 게시글의 추천 수를 가져옵니다.
+                    const countSql = 'SELECT COUNT(*) AS count FROM recommand WHERE postingid = ?;';
+                    db.get().query(countSql, postingid, (err, result) => {
+                        if (err) {
+                            console.error(err);
+                            const error = { "errorCode": "U015", "message": "추천 수를 가져오는 동안 오류가 발생했습니다." };
+                            return res.status(500).json(error);
+                        }
+
+                        const count = result[0].count;
+                        res.json({ "count": count });
+                    });
+                });
+            }
+        });
+    }
+});
+
+// 클라이언트가 '신고' 버튼을 눌렀을 때 실행되는 핸들러
+app.post('/api/report', authenticateToken, (req, res) => {
+    const { postingid, declarationreason } = req.body;
+    const userid = req.user;
+
+    // 해당 사용자와 게시글에 대한 신고가 이미 존재하는지 확인합니다.
+    const checkSql = 'SELECT * FROM declaration WHERE userid = ? AND postingid = ? LIMIT 1;';
+    db.get().query(checkSql, [userid, postingid], (err, rows) => {
+        if (err) {
+            console.error(err);
+            const error = { "errorCode": "U017", "message": "신고를 확인하는 동안 오류가 발생했습니다." };
+            return res.status(500).json(error);
+        }
+
+        if (rows.length > 0) {
+            // 이미 신고가 존재하는 경우
+            const error = { "errorCode": "U018", "message": "이미 신고한 게시글입니다." };
+            return res.status(400).json(error);
+        } else {
+            // 신고를 추가합니다.
+            const insertSql = 'INSERT INTO declaration (userid, postingid, declarationreason) VALUES (?, ?, ?);';
+            db.get().query(insertSql, [userid, postingid, declarationreason], (err) => {
+                if (err) {
+                    console.error(err);
+                    const error = { "errorCode": "U019", "message": "신고를 등록하는 동안 오류가 발생했습니다." };
+                    return res.status(500).json(error);
+                }
+
+                // 게시글의 신고 수를 확인합니다.
+                const countSql = 'SELECT COUNT(*) AS count FROM declaration WHERE postingid = ?;';
+                db.get().query(countSql, postingid, (err, result) => {
+                    if (err) {
+                        console.error(err);
+                        const error = { "errorCode": "U020", "message": "신고 수를 가져오는 동안 오류가 발생했습니다." };
+                        return res.status(500).json(error);
+                    }
+
+                    const count = result[0].count;
+
+                    // 일정 이상의 신고 수를 확인하고, 비공개로 전환합니다.
+                    const threshold = 3; // 일정 이상의 신고 수
+                    if (count >= threshold) {
+                        const updateSql = 'UPDATE posting SET disclosure = ? WHERE postingid = ?;';
+                        db.get().query(updateSql, "비공개(신고당한 게시물)", postingid, (err) => {
+                            if (err) {
+                                console.error(err);
+                                const error = { "errorCode": "U021", "message": "게시글을 비공개로 전환하는 동안 오류가 발생했습니다." };
+                                return res.status(500).json(error);
+                            }
+
+                            res.json({ "message": "게시글이 비공개로 전환되었습니다." });
+                        });
+                    } else {
+                        res.json({ "message": "신고가 접수되었습니다." });
+                    }
+                });
+            });
+        }
     });
 });
